Add files via upload
This project uses a Machine Learning algorithm to rank the prospective candidates whose profile matches with the job description based on some keywords of the job role. This project involves NLP (Natural Language Processing) to find out the 'Cosine Similarity' between the job requirement (keywords) and the candidate's profile after text preprocessing, tokenization, lemmatization, and removing 'stop words'. The higher the 'Cosine Similarity', the closer is the candidate's profile to the job role requirements. 
As a second step, one of the candidate's profile was picked as a starred candidate and the matching profiles of candidates were again found based on this candidate's profile. So, 'Cosine Similarity' was calculated between this profile and the other candidates' profiles using the same NLP algorithm.